Attribute "Max" and "Min" : BigDecimal

+++ src/test/java/junit/env/telosys/tools/generator/fakemodel/AttributeInFakeModel.java
@@ -16,6 +16,7 @@
 package junit.env.telosys.tools.generator.fakemodel;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import org.telosys.tools.commons.DatabaseUtil;
 import org.telosys.tools.commons.JavaTypeUtil;
@@ -100,8 +101,10 @@ public class AttributeInFakeModel implements Comparable<AttributeInFakeModel>, S
 	private String  _sDateAfterValue  = null ;
 	
 	//----- SPECIAL DATA for NUMERIC -----
-	private Integer _iMinValue = null ;
-	private Integer _iMaxValue = null ;
+//	private Integer _iMinValue = null ;
+//	private Integer _iMaxValue = null ;
+	private BigDecimal _iMinValue = null ;
+	private BigDecimal _iMaxValue = null ;
 	
 	//----- SPECIAL DATA for BOOLEAN -----
 	private String  _sBooleanTrueValue  = null ; // the special value for TRUE 
@@ -693,29 +696,21 @@ public class AttributeInFakeModel implements Comparable<AttributeInFakeModel>, S
 	
 	//-----------------------------------------------------------------------------
 
-//	public String getMinValue() {
-//		return _sMinValue ; 
-//	}
-	public Integer getMinValue() { // ver 3.0.0
+//	public Integer getMinValue() { // ver 3.0.0
+	public BigDecimal getMinValue() { // ver 3.0.0
 		return _iMinValue ; 
 	}
-//	public void setMinValue(String v) {
-//		_sMinValue = v ;
-//	}
-	public void setMinValue(Integer v) { // ver 3.0.0
+//	public void setMinValue(Integer v) { // ver 3.0.0
+	public void setMinValue(BigDecimal v) { // ver 3.0.0
 		_iMinValue = v ;
 	}
 	
-//	public String getMaxValue() {
-//		return _sMaxValue ; 
-//	}
-	public Integer getMaxValue() { // ver 3.0.0
+//	public Integer getMaxValue() { // ver 3.0.0
+	public BigDecimal getMaxValue() { // ver 3.0.0
 		return _iMaxValue ; 
 	}
-//	public void setMaxValue(String v) {
-//		_sMaxValue = v ;
-//	}
-	public void setMaxValue(Integer v) { // ver 3.0.0
+//	public void setMaxValue(Integer v) { // ver 3.0.0
+	public void setMaxValue(BigDecimal v) { // ver 3.0.0
 		_iMaxValue = v ;
 	}
